--- src/components/@common/FormControl/index.ts
@@ -1,1 +1,0 @@
-export { default as FormControl } from './FormControl';
