+++ src/components/@common/index.ts
@@ -1,6 +1,6 @@
 export { Box } from './Box';
 export { Header } from './Header';
 export { TextField } from './TextField';
-export { FormControl } from './FormControl';
+export { FormFieldControl } from './FormFieldControl';
 export { Button } from './Button';
 export { Tooltip } from './Tooltip';
